pyproject.toml
[project]
name = "mlops-llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.24.3",
    "pandas==2.0.3",
    "scikit-learn==1.3.0",
    "xgboost==2.0.3",
    "streamlit==1.29.0",
    "mlflow==2.9.2",
    "psycopg2-binary==2.9.9",
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pyyaml==6.0.1",
    "dvc==3.30.1",
    "dvc-s3==3.0.1",
    "matplotlib==3.7.2",
    "seaborn==0.12.2",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "dagshub>=0.6.3",
]
